Fixed method signature for type Auth to implement authgrpc.Auth

+++ internal/domain/models/account.go
@@ -10,7 +10,7 @@ type Account struct {
 	PassHash  []byte
 	Role      AccountRole
 	Status    AccountStatus
-	AppId     int64
+	AppId     int32
 }
 
 type AccountRole int32
